add auth_test.go for retrieving the access and secret keys
